refactor: extract spaCy pipeline utils into wsd/spacy_utils
Move the pipeline registry and the CPU/GPU warmup machinery out of
word_sense_disambiguation.py into a dedicated module, and fold the whole
GPU swap into one async function (swap_spacy_to_gpu) so the backend
knowledge lives in a single place instead of being split across server.py
and the wsd package.

Also address review feedback:
- run as a non-root user; COPY --chown keeps the model caches writable
  and ~/.cupy gets a home directory for the kernel cache
- default LOG_LEVEL to INFO instead of DEBUG
- drop unused DEBIAN_FRONTEND and redundant PIP_NO_CACHE_DIR
- document that the Dockerfile's model name mirrors _DEFAULT_MODEL

Verified: full swap test passes (136 requests, 0 errors, 8ms post-swap),
GPU available as non-root.
