+++ wsd/word_sense_disambiguation.py
@@ -1,10 +1,7 @@
 import logging
-import threading
 from dataclasses import dataclass
 
 import requests
-import spacy
-from thinc.api import use_ops
 
 from wsd.env import WORDNET_URL
 from wsd.letters import NOTA_LETTER_INDEX
@@ -15,6 +12,7 @@ from wsd.prompt import (
     create_marked_sentence,
     create_multiple_choice_prompt,
 )
+from wsd.spacy_utils import run_spacy_pipeline
 
 logger = logging.getLogger(__name__)
 
@@ -73,90 +71,6 @@ class DisambiguationInput:
 class WordSenseDisambiguation:
     tokens: list[DisambiguatedToken]
     entities: list[Entity]
-
-
-_PIPELINE_MODELS = {
-    'en': 'en_core_web_trf',
-    # Add more language models as needed
-}
-# language -> (pipeline, thinc backend it was built on: "numpy" or "cupy").
-# Every call must scope `use_ops(backend)`: input tensors are created on the
-# backend active in the *calling* context, which has to match the device the
-# pipeline's weights live on, and context state does not survive across
-# threads or asyncio tasks.
-_pipelines: dict[str, tuple[spacy.language.Language, str]] = {}
-_pipeline_lock = threading.Lock()
-
-
-def _load_pipeline(language: str) -> spacy.language.Language:
-    if language not in _PIPELINE_MODELS:
-        msg = f"Language '{language}' not supported"
-        raise ValueError(msg)
-    nlp = spacy.load(_PIPELINE_MODELS[language])
-    nlp.add_pipe("entityLinker", last=True)
-    return nlp
-
-
-def get_spacy_pipeline(language: str = "en"):
-    """Get the current spaCy pipeline for a language, loading it on first use.
-
-    Runs on GPU when one is available (requires cupy); falls back to CPU
-    otherwise.
-    """
-    entry = _pipelines.get(language)
-    if entry is None:
-        with _pipeline_lock:
-            entry = _pipelines.get(language)
-            if entry is None:
-                gpu_activated = spacy.prefer_gpu()
-                logger.info("spaCy GPU activated: %s", gpu_activated)
-                entry = (_load_pipeline(language), "cupy" if gpu_activated else "numpy")
-                _pipelines[language] = entry
-    return entry[0]
-
-
-def run_spacy_pipeline(text: str, language: str = "en"):
-    """Run the pipeline on a text, on the backend the pipeline was built for."""
-    get_spacy_pipeline(language)
-    nlp, backend = _pipelines[language]
-    with use_ops(backend):
-        return nlp(text)
-
-
-def warm_cpu_spacy_pipeline(language: str = "en") -> None:
-    """Load and warm a CPU pipeline (~2.5s) so requests can be served
-    immediately, without paying the GPU pipeline's one-time ~6s kernel
-    compilation first."""
-    with _pipeline_lock:
-        if language not in _pipelines:
-            with use_ops("numpy"):
-                cpu_nlp = _load_pipeline(language)
-                cpu_nlp("bank")
-            _pipelines[language] = (cpu_nlp, "numpy")
-
-
-def build_gpu_spacy_pipeline(language: str = "en"):
-    """Build and warm a GPU pipeline; returns None when no GPU is available.
-
-    Heavy (one-time ~6s CUDA kernel compilation) — run it in a worker thread
-    and publish the result with activate_spacy_pipeline.
-    """
-    if not spacy.prefer_gpu():
-        logger.info("No GPU available for spaCy; staying on CPU")
-        return None
-    gpu_nlp = _load_pipeline(language)
-    # entityLinker stays on CPU and its sqlite connection is a process-wide
-    # singleton bound to the serving thread, so it must not run here; the
-    # transformer stack is what needs the one-time kernel compilation.
-    with gpu_nlp.select_pipes(disable=["entityLinker"]):
-        gpu_nlp("bank")
-    return gpu_nlp
-
-
-def activate_spacy_pipeline(language: str, nlp) -> None:
-    """Publish a warmed GPU pipeline; subsequent requests run on it."""
-    _pipelines[language] = (nlp, "cupy")
-    logger.info("spaCy pipeline swapped to GPU")
 
 
 def _get_definitions_raw(queries: list[WordQuery], language: str = "en") -> list[list[Definition]]:
